requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
PyMuPDF==1.26.0
python-docx==1.1.2
pytesseract==0.3.13
Pillow==11.2.1
requests==2.32.3
python-dotenv==1.1.0 